Ensure handlers retrigger only when spec rematches
... so that the event of matching truly invokes the corresponding
handler(s) exactly once (unlike how it was: that handlers were being
invoked for each time a new average value matched the spec).

This is achieved by adding a flag to the "spec state" to track whether
that spec has already been determined to match the current stable color.
This flag is set when the match first occurs and is cleared when the
match first stops matching.

Also:
- in the test suite, in the `whenMatching()` section...
  - add short summaries to the so that they are easier to get the gist
    when reading.
  - reorder tests so that the features are described in the general
    order they might be used.
- gather all the methods that modify `specs` (i.e. the variable that holds
  the mapping from specs to their handlers) right next to that variable
  (don't make me scroll principle).
- use the `it()` test enclosure instead of `test()`, the former is
  slightly more readable.

Authored-by: John Ryan <jtigger@infosysengr.com>